requirements.txt
BeautifulSoup==3.2.1
Django==1.5
MySQL-python==1.2.4
Pillow==2.2.1
South==0.8.3
-e git+https://github.com/Commentor/brabeion.git@82940b8fe2e5c12a6363cec625a9831aaf0c5cdb#egg=brabeion-improved
django-analytical==0.16.0
django-appconf==0.5
django-avatar==2.0
django-blog-zinnia==0.12.3
django-forms-bootstrap==2.0.3.post1
django-jsonfield==0.8.12
django-mptt==0.5.5
django-tagging==0.3.1
django-test-exclude==0.1.0
django-waitinglist==1.0b10
django-xmlrpc==0.1.5
eventlog==0.6.2
metron==1.1
mock==1.0.1
pyparsing==1.5.7
pytz==2013.8
wsgiref==0.1.2
